+++ modulus/launch/logging/launch.py
@@ -15,11 +15,14 @@
 import re
 import sys
 import time
+from os import getcwd, makedirs
+from os.path import abspath, exists, join
 from typing import Dict, Tuple, Union
 
 import torch
 import torch.cuda.profiler as profiler
 import wandb
+from matplotlib.figure import Figure
 
 from modulus.distributed import DistributedManager, gather_loss
 
@@ -323,13 +326,50 @@ class LaunchLogger(object):
             metric_dict[step[0]] = step[1]
             wandb.log(metric_dict)
 
-    def log_mlflow_figure(self, figure, artifact_file: str = "artifact"):
+    def log_figure(
+        self,
+        figure: Figure,
+        artifact_file: str = "artifact",
+        plot_dir: str = "./",
+        log_to_file: bool = False,
+    ):
+        """Logs figures on root process to wand or mlflow. Will store it to file in case neither are selected.
+
+        Parameters
+        ----------
+        figure : Figure
+            matplotlib figure to plot
+        artifact_file : str, optional
+            File name. CAUTION overrides old files of same name
+        plot_dir : str, optional
+            output directory for plot
+        log_to_file : bool, optional
+            set to true in case figure shall be stored to file in addition to logging it to mlflow/wandb
+        """
+        dist = DistributedManager()
+        if dist.rank != 0:
+            return
+
+        if self.wandb_backend:
+            wandb.log({artifact_file: figure})
+
         if self.mlflow_backend:
             self.mlflow_client.log_figure(
                 figure=figure,
                 artifact_file=artifact_file,
                 run_id=self.mlflow_run.info.run_id,
             )
+
+        if (not self.wandb_backend) and (not self.mlflow_backend):
+            log_to_file = True
+
+        if log_to_file:
+            plot_dir = abspath(join(getcwd(), plot_dir))
+            if not exists(plot_dir):
+                makedirs(plot_dir)
+            if not artifact_file.endswith(".png"):
+                artifact_file += ".png"
+            figure.savefig(join(plot_dir, artifact_file))
 
     @classmethod
     def toggle_wandb(cls, value: bool):